Add recursive coding capability
This required reworking the way things work.  This was because the low
bit of the state is no longer the input bit value of the coder, so the
next state can't just be directly calculated by putting in the bit and
the bit can't be extracted by using the low bit.

Fixing this required splitting the convert into two arrays of size
num_states, one for a zero bit and one for a one bit.

It also required creating a next_state array for calculating the next
state.

There was also a setting to output the encoded data by symbol, not as
full bytes.  This was to simplify use in turbo decoders since they will
split the bits up.

The output order of the encoded bits was also switched.  It was really
wrong before, the first polynomial should be the first bit, the second
the second bit, and so forth.  It was backwards.  Switch it to be right.

Signed-off-by: Corey Minyard <minyard@acm.org>